+++ src/decorators/InterfaceType.ts
@@ -1,10 +1,16 @@
 import { getMetadataStorage } from "../metadata/getMetadataStorage";
 import { getNameDecoratorParams } from "../helpers/decorators";
-import { DescriptionOptions, AbstractClassOptions, ResolveTypeOptions } from "./types";
+import {
+  DescriptionOptions,
+  AbstractClassOptions,
+  ResolveTypeOptions,
+  ImplementsClassOptions,
+} from "./types";
 
 export type InterfaceTypeOptions = DescriptionOptions &
   AbstractClassOptions &
-  ResolveTypeOptions & {
+  ResolveTypeOptions &
+  ImplementsClassOptions & {
     /**
      * Set to false to prevent emitting in schema all object types
      * that implements this interface type.
@@ -20,10 +26,12 @@ export function InterfaceType(
   maybeOptions?: InterfaceTypeOptions,
 ): ClassDecorator {
   const { name, options } = getNameDecoratorParams(nameOrOptions, maybeOptions);
+  const interfaceClasses = options.implements && ([] as Function[]).concat(options.implements);
   return target => {
     getMetadataStorage().collectInterfaceMetadata({
       name: name || target.name,
       target,
+      interfaceClasses,
       autoRegisteringDisabled: options.autoRegisterImplementations === false,
       ...options,
     });