+++ src/metadata/definitions/interface-class-metadata.ts
@@ -4,4 +4,5 @@ import { TypeResolver } from "../../interfaces";
 export interface InterfaceClassMetadata extends ClassMetadata {
   resolveType?: TypeResolver<any, any>;
   autoRegisteringDisabled: boolean;
+  interfaceClasses: Function[] | undefined;
 }