+++ src/metadata/definitions/object-class-metdata.ts
@@ -1,5 +1,5 @@
 import { ClassMetadata } from "./class-metadata";
 
 export interface ObjectClassMetadata extends ClassMetadata {
-  interfaceClasses?: Function[];
+  interfaceClasses: Function[] | undefined;
 }